fix(ui): ui polish — squarify, SegmentedControl toggles, mobile toolbar, safe area, empty states
- Set --radius: 0 globally; update card, button, input to rounded-none
- Replace inline view toggles with SegmentedControl (members, tasks, templates)
- Add size prop (sm/default) to SegmentedControl
- Split timetable-client.tsx into timetable-client/ folder structure
- Fix timetable toolbar: two-row layout on mobile (nav + controls rows)
- Hide calendar empty-state overlay during tap-to-place mode on mobile
- Navbar: sticky + safe-area padding (viewport-fit=cover, env(safe-area-inset-top))
- Page-header: sticky below --safe-header-height var
- Sidebar close button: add cursor-pointer
- Calendar empty state: frosted border overlay, no rounded corners
- SimpleView + task table empty states: plain outlined box with icon
