+++ app/globals.css
@@ -8,6 +8,7 @@
   --color-background: var(--background);
   --color-foreground: var(--foreground);
   --font-sans: var(--font-sans);
+  --font-heading: var(--font-heading);
   --font-mono: var(--font-mono);
   --color-sidebar-ring: var(--sidebar-ring);
   --color-sidebar-border: var(--sidebar-border);
@@ -49,7 +50,7 @@
 
 :root {
   /* Page & surfaces */
-  --background: oklch(0.965 0.004 258); /* slight blue-gray page bg */
+  --background: oklch(1 0 0); /* white page bg */
   --foreground: oklch(0.24 0.05 258); /* dark navy text */
   --card: oklch(1 0 0);
   --card-foreground: oklch(0.24 0.05 258);
@@ -84,7 +85,7 @@
   --chart-5: oklch(0.424 0.199 265.638);
 
   /* Border radius — tighter for a clean business feel */
-  --radius: 0.375rem;
+  --radius: 0;
 
   /* Sidebar — clean white */
   --sidebar: oklch(1 0 0);
@@ -95,6 +96,9 @@
   --sidebar-accent-foreground: oklch(0.35 0.04 258);
   --sidebar-border: oklch(0.885 0.006 258);
   --sidebar-ring: oklch(0.5 0.22 263);
+
+  /* Safe area — accounts for macOS fullscreen menu bar zone */
+  --safe-header-height: calc(3.5rem + env(safe-area-inset-top, 0px));
 }
 
 .dark {
@@ -136,10 +140,13 @@
     @apply border-border outline-ring/50;
   }
   body {
-    @apply bg-background text-foreground;
+    @apply bg-background text-foreground font-sans;
   }
-  html {
-    @apply font-sans;
+  h1,
+  h2,
+  h3,
+  h4 {
+    font-family: var(--font-heading, var(--font-sans));
   }
 }
 
@@ -150,8 +157,7 @@
 }
 
 /* Subtle card elevation */
-.rounded-lg.border.bg-card,
-.rounded-xl.border.bg-card {
+.border.bg-card {
   box-shadow:
     0 1px 3px 0 oklch(0 0 0 / 0.06),
     0 1px 2px -1px oklch(0 0 0 / 0.04);
